Hotfix 2

+++ parallel/wiggletoolsIndex.py
@@ -22,12 +22,13 @@ def main():
 	cmd = ['wiggletools'] + wiggletools_args
 	run(cmd)
 
-	for match in re.finditer('write\s*(\S*.wig)\s', wiggletools_args):
-		wiggle_file = match.group(1)
-		if os.path.getsize(wiggle_file) > 0:
-			bigwig_file = re.sub('.wig$','.bw', wiggle_file)
-			run(['wigToBigWig -keepAllChromosomes -fixedSummaries',wiggle_file,chrom_lengths_file,bigwig_file])
-		os.remove(wiggle_file)
+	for i in range(1, len(wiggletools_args)):
+		if wiggletools_args[i-1] == 'write' and re.search(r'\S*.wig', wiggletools_args[i]) is not None:
+			wiggle_file = wiggletools_args[i]
+			if os.path.getsize(wiggle_file) > 0:
+				bigwig_file = re.sub('.wig$','.bw', wiggle_file)
+				run(['wigToBigWig -keepAllChromosomes -fixedSummaries',wiggle_file,chrom_lengths_file,bigwig_file])
+			os.remove(wiggle_file)
 
 if __name__ == "__main__":
 	main()
